BACKPORT: PM: sleep: Restore asynchronous device resume optimization
Before commit 7839d0078e0d ("PM: sleep: Fix possible deadlocks in core
system-wide PM code"), the resume of devices that were allowed to resume
asynchronously was scheduled before starting the resume of the other
devices, so the former did not have to wait for the latter unless
functional dependencies were present.

Commit 7839d0078e0d removed that optimization in order to address a
correctness issue, but it can be restored with the help of a new device
power management flag, so do that now.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Reviewed-by: Stanislaw Gruszka <stanislaw.gruszka@linux.intel.com>

[backport: move the new bitfield to end of the list to preserve KMI;
 hide the bitfield addition from genksyms; refresh ABI and allow break]

Bug: 502745389
Change-Id: I27c085b0a8b9275293e706633cdc5557c6659ba1
Signed-off-by: Sneh Mankad <sneh.mankad@oss.qualcomm.com>